mfd: atmel-hlcdc: implement write synchronization
Some HLCDC registers cannot be written until the previous write access has
been synchronized with the hardware. If they are written while a
synchronization is in progress, the new value (and the associated
configuration) might be silently ignored, resulting in unpredictable
behavior.

Hide the write synchronization stuff in a regmap implementation and use
this implementation instead of the generic mmio one.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>

---
Changes since v1:
 - add a 100 usec timeout