mtd: nand: brcm: kill the pdev field in brcmnand_host
The ->pdev field in the brcmnand_controller is only used to print messages
or allocate managed resources, in other words, we don't need a platform
device but the device contained by the platform_device struct.
Moreover, this information is already stored in the nand_controller struct,
so we can simply kill ->pdev and use controller->dev instead.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>

Conflicts:
	drivers/mtd/nand/brcmnand/brcmnand.c