pwm: atmel: Avoid copying driver data content
struct atmel_pwm_data is here to customize the implementation depending
on the IP version, and we plan to extend this structure with new hooks.
In order to ease the transition, store a pointer to the atmel_pwm_data
object in struct atmel_pwm_chip instead of copying its content.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>