mtd: nand: Provide several helpers to do a common NAND operations
This is part of the process of removing direct calls to ->cmdfunc()
outside of the core in order to introduce a better interface to execute
NAND operations.

Here we provide several helpers and make use of them to remove all
direct calls to ->cmdfunc(). This way, we can easily modify those
helpers to make use of the new ->exec_op() interface when available.

FIXME: Not sure we should patch all drivers to use the new helpers.
Maybe we should leave that for later when those drivers are actually
converted to the ->exec_op() interface.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>